+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # LPC Language Services Changelog
 
+## 1.1.47
+
+- Fix: [Rename (f2) reports "no project" error when file is excluded
+  #200](https://github.com/jlchmura/lpc-language-server/issues/200)
+
 ## 1.1.46
 
 - Update FluffOS efun signatures; add new efuns and variants